A simple proof of concept project showing how to save and load in Unity
This project is a small proof of concept Unity project showing how to
save and load data, specifically with mobile Android support in mind.
The project shows how to use Unity's PlayerPrefs to save small data sets
such as player preferences as well as how to serialize and read and
write data to and from binary files.